requirements.txt
streamlit==1.42.0
langchain-community==0.3.17
langchain-core==0.3.35
langchain-groq==0.2.4
langchain-google-genai==2.0.9
python-dotenv==1.0.1
docx2txt==0.8
bs4==0.0.2
pypdf==5.3.0
duckdb==0.9.2
google-generativeai==0.8.4
pysqlite3-binary
ipykernel
